+++ functions.php
@@ -132,6 +132,7 @@ get_template_part( 'lib/meta/meta-boxes-taxonomy' );
 get_template_part( 'lib/meta/meta-boxes-category-novara-live' );
 get_template_part( 'lib/meta/meta-boxes-category-tyskysour' );
 get_template_part( 'lib/meta/meta-boxes-category-dyor' );
+get_template_part( 'lib/meta/meta-boxes-post-dyor' );
 get_template_part( 'lib/meta/meta-boxes-posttype-contributor' );
 get_template_part( 'lib/meta/meta-boxes-posttype-job' );
 get_template_part( 'lib/meta/meta-boxes-posttype-newsletter' );